1) Fixed some 64 bit compatibility issues of the Python interface (which would break simple_python_example.py with GPU enabled, if compiled for 64 bit) 2) Added demo of Transmission Tomography MLEM reconstruction with large volume.

+++ Python/NiftyRec/NiftyRec/NiftyRec.py
@@ -207,15 +207,15 @@ def et_list_gpus():
 
 
 def et_is_block_multiple(integer_value): 
-    L.et_array_is_block_multiple.restype  = c_int
-    L.et_array_is_block_multiple.argtypes = [c_int]
+    L.et_array_is_block_multiple.restype  = c_ushort
+    L.et_array_is_block_multiple.argtypes = [c_ushort]
     integer_value = int32(integer_value)
     status = L.et_array_is_block_multiple(integer_value)
     return status 
 
 
 def et_get_block_size():
-    L.et_array_get_block_size.restype = c_int
+    L.et_array_get_block_size.restype = c_ushort
     L.et_array_get_block_size.argtypes = []
     return L.et_array_get_block_size()
